requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
langchain-groq
langchain-core
langchain-experimental
pydantic==1.10.13  # LangChain compatibility
neo4j
pdfplumber
python-dotenv
tiktoken
rank_bm25 
fitz 
tools 
deepeval 
pypdf 
duckduckgo-search